% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hakaiApi-package.R
\docType{package}
\name{hakaiApi-package}
\alias{hakaiApi}
\alias{hakaiApi-package}
\title{hakaiApi: Authenticated HTTP Request Client for the 'Hakai' API}
\description{
Initializes a class that obtains API credentials and provides a method to use those credentials to make GET requests to the 'Hakai' API server. Usage instructions are documented at \url{https://hakaiinstitute.github.io/hakai-api/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HakaiInstitute/hakai-api-client-r}
  \item Report bugs at \url{https://github.com/HakaiInstitute/hakai-api-client-r/issues}
}

}
\author{
\strong{Maintainer}: Taylor Denouden \email{taylor.denouden@hakai.org}

Authors:
\itemize{
  \item Brett Johnson \email{brett.johnson@hakai.org}
}

Other contributors:
\itemize{
  \item Nate Rosenstock \email{nate.rosenstock@hakai.org} [contributor]
  \item Sam Albers \email{sam.albers@hakai.org} [contributor]
  \item Chris Davis \email{chris.davis@hakai.org} [contributor]
}

}
\keyword{internal}
