% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Client.R
\docType{data}
\name{Client}
\alias{Client}
\title{A Hakai API Client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Client
}
\arguments{
\item{api_root}{The root url for all API Requests. Defaults to https://hecate.hakai.org/api}
}
\description{
This class allows you to make authenticated API requests
}
\examples{
Client$new()
}
\keyword{hakai}
