% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Client.R
\name{Client}
\alias{Client}
\title{A Hakai API Client}
\description{
This class allows you to make authenticated API requests
}
\examples{
Client$new()
}
\keyword{hakai}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Client$new()}}
\item \href{#method-get}{\code{Client$get()}}
\item \href{#method-remove_old_credentials}{\code{Client$remove_old_credentials()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$new(api_root = "https://hecate.hakai.org/api")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_root}}{The root url for all API Requests. Defaults to https://hecate.hakai.org/api}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$get(endpointUrl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_old_credentials"></a>}}
\subsection{Method \code{remove_old_credentials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$remove_old_credentials()}\if{html}{\out{</div>}}
}

}
}
